1 timesries endpoint
