+++ backend/app/api/routes/v1/__init__.py
@@ -12,7 +12,7 @@ from .oauth import router as oauth_router
 from .summaries import router as summaries_router
 from .suunto_debug import router as suunto_debug_router
 from .sync_data import router as sync_data_router
-from .time_series import router as time_series_router
+from .timeseries import router as timeseries_router
 from .users import router as users_router
 from .vendor_workouts import router as vendor_workouts_router
 from .workouts import router as workouts_router
@@ -38,7 +38,7 @@ v1_router.include_router(dashboard_router, prefix="/dashboard", tags=["dashboard
 
 # RFC Taxonomy Routes
 v1_router.include_router(summaries_router, tags=["Summaries"])
-v1_router.include_router(time_series_router, tags=["Timeseries"])
+v1_router.include_router(timeseries_router, tags=["Timeseries"])
 v1_router.include_router(events_router, tags=["Events"])
 
 __all__ = ["v1_router"]
