+++ backend/app/api/routes/v1/timeseries.py
@@ -1,0 +1,40 @@
+from datetime import datetime
+from typing import Annotated, Literal
+from uuid import UUID
+
+from fastapi import APIRouter, Query
+
+from app.database import DbSession
+from app.schemas.common_types import PaginatedResponse
+from app.schemas.series_types import SeriesType
+from app.schemas.timeseries import (
+    BloodGlucoseSample,
+    HeartRateSample,
+    HrvSample,
+    Spo2Sample,
+    StepsSample,
+    TimeSeriesQueryParams,
+)
+from app.services import ApiKeyDep, timeseries_service
+
+router = APIRouter()
+
+
+@router.get("/users/{user_id}/timeseries")
+async def get_timeseries(
+    user_id: UUID,
+    type: SeriesType,
+    start_time: str,
+    end_time: str,
+    db: DbSession,
+    _api_key: ApiKeyDep,
+    resolution: Literal["raw", "1min", "5min", "15min", "1hour"] = "raw",
+    cursor: str | None = None,
+    limit: Annotated[int, Query(ge=1, le=100)] = 50,
+) -> PaginatedResponse[HeartRateSample | HrvSample | Spo2Sample | BloodGlucoseSample | StepsSample]:
+    """Returns granular time series data (biometrics or activity)."""
+    params = TimeSeriesQueryParams(
+        start_datetime=datetime.fromisoformat(start_time),
+        end_datetime=datetime.fromisoformat(end_time),
+    )
+    return await timeseries_service.get_timeseries(db, user_id, type, params)
