+++ backend/app/integrations/celery/tasks/process_upload_task.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from sqlalchemy.orm import Session
 
 from app.database import SessionLocal
-from app.services import event_record_service, time_series_service
+from app.services import event_record_service, timeseries_service
 from app.services.apple.apple_xml.aws_service import s3_client
 from app.services.apple.apple_xml.xml_service import XMLService
 from celery import shared_task
@@ -69,6 +69,6 @@ def _import_xml_data(db: Session, xml_path: str, user_id: str) -> None:
             event_record_service.create(db, record)
             event_record_service.create_detail(db, detail)
         if heart_rate_records:
-            time_series_service.bulk_create_samples(db, heart_rate_records)
+            timeseries_service.bulk_create_samples(db, heart_rate_records)
         if step_records:
-            time_series_service.bulk_create_samples(db, step_records)
+            timeseries_service.bulk_create_samples(db, step_records)
