+++ backend/app/schemas/__init__.py
@@ -121,23 +121,21 @@ from .system_info import (
     DataPointsInfo,
     SystemInfoResponse,
 )
-from .time_series import (
+from .timeseries import (
+    BloodGlucoseSample,
+    HeartRateSample,
     HeartRateSampleCreate,
     HeartRateSampleResponse,
+    HrvSample,
+    SleepStageSample,
+    Spo2Sample,
     StepSampleCreate,
     StepSampleResponse,
+    StepsSample,
     TimeSeriesQueryParams,
     TimeSeriesSampleCreate,
     TimeSeriesSampleResponse,
     TimeSeriesSampleUpdate,
-)
-from .timeseries_samples import (
-    BloodGlucoseSample,
-    HeartRateSample,
-    HrvSample,
-    SleepStageSample,
-    Spo2Sample,
-    StepsSample,
 )
 from .user import (
     UserCreate,
